fix: renamed Repository files and updated Controllers files
