+++ backend/src/repository/userRestrictionRepository.js
@@ -1,5 +1,5 @@
 const { databaseQuery } = require('../config/db');
-const { getRestrictionByCod } = require('./restrictionModel')
+const { getRestrictionByCod } = require('./restrictionRepository')
 
 const getUserRestriction = async (cod_usuario = null)=>{
     try{
